refactor: split the remaining ported landing bodies into sections
HowItWorksBody.tsx (340 lines), PrivacyBody.tsx (327) and CreatureBody.tsx
(262) were the last oversized files in apps/landing - all three pure static
JSX ported from hand-authored HTML with no hooks, handlers or logic.

They split verbatim into sections/ directories beside each body, cutting at
the markup's own structural units: how-it-works at its <aside> and #view-*
sections, privacy at its <h2>-delimited blocks, creature at doodles, stage
and overlays. The shells are 38, 25 and 24 lines.

No JSX was rewritten, matching the approach used for the home files in #917.
Verified against pre-split builds: tag sequence with ids and classes and
whitespace-collapsed visible text are both byte-identical on all three pages
(189/23474, 191/15134, 94/10560).

The inline-hex eslint exemption gains creature/sections/**, scoped rather
than a creature/** glob so the directory stays covered by the brand-token
rule if a hand-written file lands there. All 14 new files are added to
SECRET_SCAN_FILES, which fails open for anything unlisted.

Provenance headers on the privacy sections no longer claim privacy-0.js
binds to the page; it is a console easter egg with no DOM binding, and
check-parity is what actually pins that page. Corrected on PrivacyBody.tsx
too, where the claim originated.
