+++ scripts/check-landing-drift.mjs
@@ -71,10 +71,24 @@ const SECRET_SCAN_FILES = [
   'apps/landing/src/components/home/beats/Beat3.tsx',
   'apps/landing/src/components/home/beats/Beat4.tsx',
   'apps/landing/src/components/creature/CreatureBody.tsx',
+  'apps/landing/src/components/creature/sections/Doodles.tsx',
+  'apps/landing/src/components/creature/sections/Stage.tsx',
+  'apps/landing/src/components/creature/sections/Overlays.tsx',
   'apps/landing/src/components/download/DownloadBody.tsx',
   'apps/landing/src/components/download/DownloadCards.tsx',
   'apps/landing/src/components/how-it-works/HowItWorksBody.tsx',
+  'apps/landing/src/components/how-it-works/sections/Sidebar.tsx',
+  'apps/landing/src/components/how-it-works/sections/Overview.tsx',
+  'apps/landing/src/components/how-it-works/sections/Boot.tsx',
+  'apps/landing/src/components/how-it-works/sections/Flows.tsx',
+  'apps/landing/src/components/how-it-works/sections/IpcReference.tsx',
+  'apps/landing/src/components/how-it-works/sections/Subsystems.tsx',
+  'apps/landing/src/components/how-it-works/sections/CheatSheet.tsx',
   'apps/landing/src/components/privacy/PrivacyBody.tsx',
+  'apps/landing/src/components/privacy/sections/IntroShort.tsx',
+  'apps/landing/src/components/privacy/sections/Extension.tsx',
+  'apps/landing/src/components/privacy/sections/Desktop.tsx',
+  'apps/landing/src/components/privacy/sections/Footer.tsx',
   'apps/landing/src/components/SiteFooter.tsx',
   'apps/landing/src/components/BackLink.tsx',
   'apps/landing/src/lib/site-links.ts',
